+++ tests/cpp-tests/Source/NewAudioEngineTest/NewAudioEngineTest.cpp
@@ -483,8 +483,8 @@ bool AudioWavTest::init()
             {
                 AudioEngine::stop(_audioID);
                 _audioID = AudioEngine::play2d(_wavFiles[--_curIndex]);
-                _stateLabel->setString(StringUtils::format("[index: %d] %s", _curIndex,
-                                                           FileUtils::getFileShortName(_wavFiles[_curIndex]).c_str()));
+                _stateLabel->setString(fmt::format("[index: {}] {}", _curIndex,
+                                                           FileUtils::getFileShortName(_wavFiles[_curIndex])));
             }
         });
         playPrev->setPosition(layerSize.width * 0.35f, layerSize.height * 0.5f);
@@ -495,8 +495,8 @@ bool AudioWavTest::init()
             {
                 AudioEngine::stop(_audioID);
                 _audioID = AudioEngine::play2d(_wavFiles[++_curIndex]);
-                _stateLabel->setString(StringUtils::format("[index: %d] %s", _curIndex,
-                                                           FileUtils::getFileShortName(_wavFiles[_curIndex]).c_str()));
+                _stateLabel->setString(fmt::format("[index: {}] {}", _curIndex,
+                                                           FileUtils::getFileShortName(_wavFiles[_curIndex])));
             }
         });
         playNext->setPosition(layerSize.width * 0.65f, layerSize.height * 0.5f);
@@ -516,8 +516,8 @@ void AudioWavTest::onEnter()
     {
         _curIndex = 0;
         _audioID  = AudioEngine::play2d(_wavFiles[_curIndex]);
-        _stateLabel->setString(StringUtils::format("[index: %d] %s", _curIndex,
-                                                   FileUtils::getFileShortName(_wavFiles[_curIndex]).c_str()));
+        _stateLabel->setString(fmt::format("[index: {}] {}", _curIndex,
+                                                   FileUtils::getFileShortName(_wavFiles[_curIndex])));
     }
 }